Add GNU GPL free software license
This software is free. The GNU General Public License is intended to
guarantee your freedom to
share and change all versions of a program--to make sure it remains free
software for all its users. Specifically, free software means users have
the four essential freedoms: (0) to run the program, (1) to study and
change the program in source code form, (2) to redistribute exact
copies, and (3) to distribute modified versions.