Add help menu
Using Microsoft translations:
Online help:
A capability of many programs and operating systems to display advice or instructions for using their features when so requested by the user, as by a screen button or menu item or a function key.
Support:
The menu item that takes users to a site that helps them troubleshoot, diagnose and fix problems, search help articles, and communicate directly with support.
Donate to support project development:
Dynamics 365 for Sales, Field Service, Customer Service, and Project Service